refactor: 핀치줌/팬 제스처를 useSheetZoomPan 훅으로 추출 (순환 의존은 ref+isZoomActive/resetZoom으로 해소)
